+++ src/PortProcedures.h
@@ -36,8 +36,8 @@
 
 namespace scheme {
 
-    ucs4string utf8ToUtf32(const char* s, int len);
-    ucs4string utf16ToUtf32(const char* s, int len);
+    ucs4string utf8ToUtf32(const char* s, size_t len);
+    ucs4string utf16ToUtf32(const char* s, size_t len);
     char* utf32toUtf8(const ucs4string& s);
     Object makeCustomTextualInputOutputPortEx(VM* theVM, int argc, const Object* argv);
     Object makeCustomBinaryInputOutputPortEx(VM* theVM, int argc, const Object* argv);
